+++ sacred/config_scope.py
@@ -116,6 +116,7 @@ class ConfigScope(dict):
         self._initialized = False
         self.added_values = set()
         self.typechanges = {}
+        self.ignored_fallback_writes = []
 
     def __call__(self, fixed=None, preset=None, fallback=None):
         """
@@ -162,6 +163,7 @@ class ConfigScope(dict):
         eval(self._body_code, copy(self._func.__globals__), cfg_locals)
         self.added_values = cfg_locals.revelation()
         self.typechanges = cfg_locals.typechanges
+        self.ignored_fallback_writes = cfg_locals.ignored_fallback_writes
 
         # fill in the unused presets
         recursive_fill_in(cfg_locals, preset)